Инфраструктура для машинного обучения #43 fill-empty schema for huge objects

+++ core/src/main/java/ru/hypernavi/core/classify/scheme/Picture.java
@@ -3,13 +3,12 @@ package ru.hypernavi.core.classify.scheme;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.ByteArrayInputStream;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
+import java.util.Objects;
 
 
 import org.apache.commons.io.FileUtils;
@@ -18,25 +17,25 @@ import org.apache.commons.logging.LogFactory;
 import org.openimaj.image.ImageUtilities;
 import org.openimaj.image.MBFImage;
 import ru.hypernavi.commons.Chain;
+import ru.hypernavi.ml.HugeObject;
 import ru.hypernavi.util.MD5;
 import ru.hypernavi.util.MoreIOUtils;
 
 /**
  * Created by amosov-f on 03.09.15.
  */
-public final class Picture {
+public final class Picture implements HugeObject {
     private static final Log LOG = LogFactory.getLog(Picture.class);
 
     @NotNull
     private final URL url;
-    @NotNull
-    private final MBFImage image;
+    @Nullable
+    private MBFImage image;
     @Nullable
     private Chain chain;
 
-    public Picture(@NotNull final URL url, @NotNull final MBFImage image) {
+    private Picture(@NotNull final URL url) {
         this.url = url;
-        this.image = image;
     }
 
     @NotNull
@@ -46,7 +45,7 @@ public final class Picture {
 
     @NotNull
     public MBFImage getImage() {
-        return image;
+        return Objects.requireNonNull(image);
     }
 
     @Nullable
@@ -54,7 +53,21 @@ public final class Picture {
         return chain;
     }
 
-    public void setChain(@Nullable final Chain chain) {
+    @Override
+    public void fill() {
+        try {
+            image = cachedImage(url.toString());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public void empty() {
+        image = null;
+    }
+
+    private void setChain(@Nullable final Chain chain) {
         this.chain = chain;
     }
 
@@ -70,45 +83,50 @@ public final class Picture {
 
     @Nullable
     public static Picture download(@NotNull final String url) {
-        final byte[] imageBytes = getOrDownload(url);
-        if (imageBytes == null) {
+        if (!cache(url)) {
             return null;
         }
+        final Picture picture;
         try {
-            final Picture picture = new Picture(new URL(url), ImageUtilities.readMBF(new ByteArrayInputStream(imageBytes)));
-            LOG.info("Image loaded: " + url);
-            return picture;
+            picture = new Picture(new URL(url));
         } catch (MalformedURLException e) {
-            // catches in getOrDownload
+            // catches in cache()
             throw new RuntimeException(e);
+        }
+        // checking image loading
+//        try {
+//            cachedImage(url);
+//        } catch (IOException e) {
+//            LOG.error("Can't read MBF image!", e);
+//            return null;
+//        }
+        LOG.info("Picture loaded: " + url);
+        return picture;
+    }
+
+    private static boolean cache(@NotNull final String url) {
+        if (cacheFile(url).exists()) {
+            return true;
+        }
+        LOG.info("Downloading image: " + url);
+        final byte[] data;
+        try {
+            data = MoreIOUtils.read(url);
         } catch (IOException e) {
-            LOG.error("Can't read MBF image!", e);
-            return null;
+            LOG.error("No image by url: " + url, e);
+            return false;
+        }
+        try {
+            FileUtils.writeByteArrayToFile(cacheFile(url), data);
+            return true;
+        } catch (IOException e) {
+            throw new RuntimeException(e);
         }
     }
 
-    @Nullable
-    private static byte[] getOrDownload(@NotNull final String url) {
-        try {
-            return FileUtils.readFileToByteArray(cacheFile(url));
-        } catch (FileNotFoundException ignored) {
-            LOG.info("Downloading image: " + url);
-            final byte[] data;
-            try {
-                data = MoreIOUtils.read(url);
-            } catch (IOException e) {
-                LOG.error("No image by url: " + url, e);
-                return null;
-            }
-            try {
-                FileUtils.writeByteArrayToFile(cacheFile(url), data);
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-            return data;
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
+    @NotNull
+    private static MBFImage cachedImage(@NotNull final String url) throws IOException {
+        return ImageUtilities.readMBF(cacheFile(url));
     }
 
     @NotNull
